refactor(ai-tools): move citationIds into the ai utils bucket
`ai/tools/` holds tool cores plus their adapters; citationIds is a two-function
helper, and its home is the AI subsystem's utils bucket — where `textPaging.ts`
already sits despite being consumed by a tool core.
